Add files via upload
This is a project that I did in our Introduction to Secure Software Development course. It is based on the Java Vulnerable Lab that can be downloaded from github. The vulnerability let's an attacker use SQL injection to gain access to log into the JVL system using string concatenation.  That vulnerability is fixed in the project. 
You can deploy the WAR file to your Java IDE and access the file. Accompanying the WAR file is a PDF file which is describes steps taken to fix the vulnerability.